Client only and working!

+++ Server/MatchResult.cs
@@ -1,6 +1,6 @@
 using System;
 public class MatchResult {
-    public float matchTime;
+    public int matchTime;
     public int cardIndex;
 
 }
